Back Finished

+++ App/Controller/EnvoyeurController.php
@@ -4,21 +4,40 @@ require_once dirname(__DIR__, 2) . DIRECTORY_SEPARATOR . "connection.php";
 class EnvoyeurController{
 
     //CREATION
-    public function creationEnvoyeur (int $idEnvoi, string $nomEnvoyeur, string $emailEnvoyeur, $frais_Envoi, $dateEnvois, $nomRecepteur, $contactRecepteur) {
-
-        //$data = file_get_contents("php://input");
-        //$data_ = json_decode($data);
-
+    public function creationEnvoyeur () {
+        $data = file_get_contents("php://input");
+        $data_ = json_decode($data, true);
         $db = connectDatabase();
         $result = [];
-        
 
-        $query = "INSERT INTO envoyeur (idEnvoi,nomEnvoyeur,emailEnvoyeur,frais_Envoi,dateEnvois,nomRecepteur,contactRecepteur)
-        VALUES('$idEnvoi', '$nomEnvoyeur', '$emailEnvoyeur', '$frais_Envoi', '$dateEnvois', '$nomRecepteur', '$contactRecepteur');";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
+        if (!isset($data_['idEnvoi'], $data_['nomEnvoyeur'], $data_['emailEnvoyeur'], $data_['frais_Envoi'], $data_['dateEnvois'], $data_['nomRecepteur'], $data_['contactRecepteur'])){
+            $result['status'] = 'error';
+            $result['message'] = "argument missmatch";
+        } else {
+            $idEnvoi = $data_['idEnvoi'];
+            $nomEnvoyeur = $data_['nomEnvoyeur'];
+            $emailEnvoyeur = $data_['emailEnvoyeur'];
+            $frais_Envoi = $data_['frais_Envoi'];
+            $dateEnvois = $data_['dateEnvois'];
+            $nomRecepteur = $data_['nomRecepteur'];
+            $contactRecepteur = $data_['contactRecepteur'];
 
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            $query = "INSERT INTO envoyeur (idEnvoi,nomEnvoyeur,emailEnvoyeur,frais_Envoi,dateEnvois,nomRecepteur,contactRecepteur)
+            VALUES('$idEnvoi', '$nomEnvoyeur', '$emailEnvoyeur', '$frais_Envoi', '$dateEnvois', '$nomRecepteur', '$contactRecepteur');";
+    
+            try {
+                $stmt = $db->prepare($query);
+                $stmt->execute();
+
+                $result['status'] = "success";
+                $result['message'] = "sender insertion successfull";
+
+            } catch (PDOException $e) {
+                $result['status'] = "error";
+                $result['message'] = "an erro occurred while inserting a new sender";
+            }
+    
+        }
 
         return json_encode($result);
     }
@@ -28,40 +47,79 @@ class EnvoyeurController{
         $db = connectDatabase();
         $result = [];
 
-        $query = "SELECT * FROM envoyeur";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
-
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        try{
+            $query = "SELECT * FROM envoyeur";
+            $stmt = $db->prepare($query);
+            $stmt->execute();
+    
+            $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        } catch (PDOException $e){
+            $result['status'] = "error";
+            $result['message'] = "an error occured while listing senders";
+        }
 
         return json_encode($result);
     }
 
     //UPDATE
-    public function updateEnvoyeur(string $newNom, $oldNom) {
+    public function updateEnvoyeur() {
         $db = connectDatabase();
         $result = [];
 
-        $query = "UPDATE envoyeur SET nomEnvoyeur = '". $newNom ."' WHERE nomEnvoyeur = '". $oldNom ."';";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
+        $data = file_get_contents("php://input");
+        $data_ = json_decode($data, true);
 
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        if (!isset($data_['newNom'], $data_['oldNom'])){
+            $result['status'] = "error";
+            $result['message'] = "arguments missing";
+
+        } else {
+            $newNom = $data_['newNom'];
+            $oldNom = $data_['oldNom'];
+
+            $query = "UPDATE envoyeur SET nomEnvoyeur = '". $newNom ."' WHERE nomEnvoyeur = '". $oldNom ."';";
+
+            try {
+                $stmt = $db->prepare($query);
+                $stmt->execute();
+
+                $result['status'] = "success";
+                $result['message'] = "sender modification successfull";
+
+            } catch (PDOException $e){
+                $result['status'] = "error";
+                $result['message'] = "an error occured while updating a sender's information";
+            }
+        }
 
         return json_encode($result);
     }
 
     //SUPPRIMER
-    public function deleteEnvoyeur(string $delete) {
+    public function deleteEnvoyeur() {
         $db = connectDatabase();
         $result = [];
+        $data = file_get_contents("php://input");
+        $data_ = json_decode($data, true);
 
-        $query = "DELETE FROM envoyeur WHERE idEnvoyeur = '". $delete ."';";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
+        if (!isset($data_['delete'])){
+            $result['status'] = "error";
+            $result['message'] = "sender id not specified";
+        } else {
+            $delete = $data_['delete'];
+            $query = "DELETE FROM envoyeur WHERE idEnvoyeur = '". $delete ."';";
 
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            try{
+                $stmt = $db->prepare($query);
+                $stmt->execute();
+                $result['status'] = "success";
+                $result['message'] = "sender deleted successfully";
 
+            } catch (PDOException $e){
+                $result['status'] = "error";
+                $result['message'] = "an erro occured while trying to delete a sender";
+            }
+        }
         return json_encode($result);
     }
 }
