+++ App/Controller/ItineraireController.php
@@ -3,41 +3,36 @@ require_once dirname(__DIR__, 2) . DIRECTORY_SEPARATOR . "connection.php";
 
 class ItineraireController {
 
-    /*public function NouveauItineraire(string $search) {
-        
-        //$data = file_get_contents("php://input");
-        //$data_ = json_decode($data);
-
-        $db = connectDatabase();
-        $result = [];
-
-        $query = "SELECT * FROM itineraire WHERE villArr = '" . $search . "' 
-            OR villDep = '" . $search . "' 
-            OR codeIt = '" . $search ."';";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
-
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
-
-        return json_encode($result);
-    }*/
-
     //CREATION
-    public function creationItineraire (string $codeIt, string $villDep, string $villArr) {
-
-        //$data = file_get_contents("php://input");
-        //$data_ = json_decode($data);
+    public function creationItineraire () {
+        $data = file_get_contents("php://input");
+        $data_ = json_decode($data, true);
 
         $db = connectDatabase();
         $result = [];
-        
 
-        $query = "INSERT INTO itineraire (codeIt,villDep,villArr)
-        VALUES('$codeIt', '$villDep', '$villArr');";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
+        if (!isset($data_['codeIt'], $data_['villDep'], $data_['villAr'])){
+            $result['status'] = "error";
+            $result['message'] = "arguments not set";
+        } else {
+            $codeIt = $data_['codeIt'];
+            $villDep = $data_['villeDep'];
+            $villArr = $data_['villeAr'];
+            $query = "INSERT INTO itineraire (codeIt,villDep,villArr)
+            VALUES('$codeIt', '$villDep', '$villArr');";
 
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            try {
+                $stmt = $db->prepare($query);
+                $stmt->execute();
+
+                $result['status'] = "success";
+                $result['message'] = "insertion successfull";
+
+            } catch (PDOException $e){
+                $result['status'] = "error";
+                $result['message'] = "an error occured while trying to do an insertion";
+            }
+        }
 
         return json_encode($result);
     }
@@ -57,24 +52,48 @@ class ItineraireController {
             $query = "SELECT * FROM itineraire;";
         }
 
-        $stmt = $db->prepare($query);
-        $stmt->execute();
+        try {
+            $stmt = $db->prepare($query);
+            $stmt->execute();
+            $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        } catch (PDOException $e) {
+            $result['status'] = "error";
+            $result['message'] = "an error occured while trying to do the listing";
+        }
 
         return json_encode($result);
     }
 
     //UPDATE
-    public function updateItineraire(string $update) {
+    public function updateItineraire() {
         $db = connectDatabase();
         $result = [];
+        $data = file_get_contents("php://input");
+        $data_ = json_decode($data, true);
 
-        $query = "UPDATE itineraire SET villArr = '". $update ."' WHERE villArr = 'TOAMASINA';";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
+        if (!isset($data_['villArr'], $data_['villDep'], $data_['codeIt'])){
+            $result['status'] = "error";
+            $result['message'] = "arguments missing";
+        } else {
+            $villAr = $data_['villArr'];
+            $villDep = $data_['villDep'];
+            $codeIt = $data_['codeIt'];
 
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            $query = "UPDATE itineraire SET villArr = '$villAr', villDep = '$villDep' WHERE codeIt = '$codeIt';";
+            
+            try {
+                $stmt = $db->prepare($query);
+                $stmt->execute();
+
+                $result['status'] = "success";
+                $result['message'] = "update succcessfull";
+
+            } catch (PDOException $e){
+                $result['status'] = "error";
+                $result['message'] = "an erro occured while updating informations";
+            }
+        }
 
         return json_encode($result);
     }
@@ -83,16 +102,29 @@ class ItineraireController {
     public function deleteItineraire(string $delete) {
         $db = connectDatabase();
         $result = [];
+        $data = file_get_contents("php://input");
+        $data_ = json_decode($data, true);
 
-        $query = "DELETE FROM itineraire WHERE codeIt = '". $delete ."';";
-        $stmt = $db->prepare($query);
-        $stmt->execute();
+        if (!isset($data_['delete'])){
+            $result['status'] = "error";
+            $result['message'] = "missing argument";
+        } else {
+            $delete = $data_['delete'];
 
-        $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            $query = "DELETE FROM itineraire WHERE codeIt = '". $delete ."';";
+            try {
+                $stmt = $db->prepare($query);
+                $stmt->execute();
+                $result['status'] = "success";
+                $result['message'] = "deletion successfull";
+
+            } catch (PDOException $e) {
+                $result['status'] = "error";
+                $result['message'] = "an error occured while trying a deletion";
+            }
+        }
 
         return json_encode($result);
     }
-
-
 
 }
